+++ lib/twilio-ruby/rest/supersim/v1/fleet.rb
@@ -377,7 +377,7 @@ module Twilio
                     end
                     
                     ##
-                    # @return [String] The total data usage (download and upload combined) in Megabytes that each Super SIM assigned to the Fleet can consume during a billing period (normally one month). Value must be between 1MB (1) and 2TB (2,000,000). Defaults to 1GB (1,000).
+                    # @return [String] The total data usage (download and upload combined) in Megabytes that each Super SIM assigned to the Fleet can consume during a billing period (normally one month). Value must be between 1MB (1) and 2TB (2,000,000). Defaults to 250MB.
                     def data_limit
                         @properties['data_limit']
                     end
@@ -389,7 +389,7 @@ module Twilio
                     end
                     
                     ##
-                    # @return [Boolean] Defines whether SIMs in the Fleet are capable of sending and receiving machine-to-machine SMS via Commands. Defaults to `true`.
+                    # @return [Boolean] Defines whether SIMs in the Fleet are capable of sending and receiving machine-to-machine SMS via Commands. Defaults to `false`.
                     def sms_commands_enabled
                         @properties['sms_commands_enabled']
                     end